Initial commit with version 3.2.4 from SVN
Forkes the current snapshot of PECReaderFwk to GitHub. Most of the snapshot has been tagged in SVN with release version 3.2.4 [1]; however, on top of it the commit contains a plugin to calculate single-top-related observables. This is anlalysis-specific code, and eventually it will be excluded from the framework. But for the time being it is convenient to host the analysis plugins with the framework.
[1] https://svnweb.cern.ch/trac/singletop/browser/PECLocalFwk/tags/3.2.4