Add access to generator-level jets
The code is updated to match the (forthcoming) release 53X.02.05 of the
single-top package. The package can store generator-level jets as
four-momenta only or add a number of status-2 heavy-flavour quarks
nearby, depending on the settings. Currently, PECFwk reads four-momenta
only; although, code blocks to access the numbers of heavy flavours is
included but commented out. The GenJet class contains fields to host
numbers of quarks nearby.