Trigger selection was factorised out
The interface for the trigger selection was changed to hide all the access to the trigger tree in the dedicated classes. Now PECReader only gets the trigger tree from a PEC file and deletes it when the file is closed. Events from the tree will be read in a class derieved from TriggerSelectionInterface. This change also prepares to the switch to "slim" trigger results in the new version of PEC files.