TriggerSelectionInterface has been improved
Methods are more logical now: ReadNextEvent is given the ID of the next event in the tree as this information is needed to decide which branches to read. The pointer to the trigger tree has been added as a protected member along with the corresponding event counters; now they are a part of interface. Meaningful constructor are defined. Certain methods turned to non-constant as it suits their nature better. The documentation improved.

Note that this commit introduces compilation errors.