Guard logic of WegihtBTag against boundary cases
Make the class return a meaningful weight (not isnan) at any
circumstances, even if the b-tagging efficiency is 0 or 1 for a b-tagged
or untagged jet. However, these should be very rare cases. In addition,
optimise the computations.

It has been checked that this commit does not alter shape of
distributions over b-tag event weight.