Modify the Makefile to be more flexible
Tried to put object files in subdirectories in each module (e.g., put
them in core/obj/, external/JEC/obj/, etc.) but failed. Apparently,
there is only a very limited number of ways to build names of
dependencies from names of targets, and it does not allow the desired
approach with a single Makefile.