Add text files with JEC and their uncertainties
The files with JEC are taken from database uploads cited in tables in
[1]. Uncertainties are taken from [2], which is also referenced in [1].
Only relevant files are kept, which means that all jet types but
AK5PFchs are dropped and L5 corrections and JEC uncertainty sources are
not included. L2, L3, and L2L3Residual corrections for data and MC are
identical, and the files are kept for simplicity of the setup. L1FastJet
corrections differ as expected. Files with uncertainties are not
identical as well, but I do not understand meaning of the uncertainties
for real data (they are supposed to be evaluated for MC only); could it
be inverted uncertainties?

[1] https://twiki.cern.ch/twiki/bin/viewauth/CMS/JECDataMC#Jet_Energy_Corrections
[2] https://twiki.cern.ch/twiki/bin/viewauth/CMS/JECUncertaintySources#2012_JEC