Change formula for JER to match the definition in CMSSW_5_3_11
Validated JEC and JER on the fly against the factors stored in PEC
samples (53X.03.00). JEC factors agree perfectly. JER factors match
exactly in ~80% of jets, while the other ~20% reveal a discrepancy
without any obvious pattern. It might be because the SmearedJetProducerT
plugin in CMSSW deploys a complex metrics to match to generator-level
jets. In addition, unmatched jets are not smeared at all, but their
relative number is small, especially in ttbar datasets.