In Makefiles remove defaults for Boost installation at lxplus
Environment for Boost libraries must be set by dedicated scripts or by
the user, not in Makefiles.

It seems that now in AFS installations there are no nasty postfixes in
the Boost include directory and in all libraries. Thus in future could
drop additional variables and use only BOOST_ROOT.