Drop building of a static library
From now on the framework can only be used as a shared library. I put
"4.0" as the version number of the library since the next target release
is 4.0.0, but most like no backward compatibility will be provided
within a major release cycle, and the user will be required to update
the framework whenever a new version is released. In this case library
version does not have any practical meaning.

This fixes #35.