Update paths to PEC files in test programs
Files are located at T2_FR_IPNL and are not available for general
public, but it should be fine for the time being. Will publish some
small files for the next full production release.

Program "minimal" still fails with a segfault in PECReader::Initialize,
so the problem is not in the missing input files. Will need to
investigate. Program "multithread" works fine apart from complains about
missing branches with jet raw momentum and area.