Reorganize source files into modules
Now source files are grouped into modules, which are compiled into
independent static libraries. The final shared library is built out of
these static ones. This is needed because there are namesake source
files in different modules, and thus the corresponding object files
cannot be stored in a single directory.

For the moment, external packages are not built. Will leave this as is
for some time to simplify modifications to the code.