Remove ClassDef macro in pec::EventID
The test runs fine without it, so it seems that the ClassDef macro is
somehow not needed when a ROOT dictionary is created in the
genreflex-like way. ROOT manual [1] says the macro is only required for
classes inheriting from TObject and when ROOT i/o is needed. Same is
suggested here [2]. Obviously, PEC classes need to be read from a file,
which should imply that ClassDef is needed. At the same time it is not
used in most data format classes in CMSSW (including all PAT classes)
and is not mentioned in the corresponding documentation [3]. Maybe, the
build tool adds it automatically...

Will try classes without the macro. If this does not work, will put it
in.

[1] https://root.cern.ch/root/htmldoc/guides/users-guide/ROOTUsersGuide.html#adding-a-class-with-a-shared-library
[2] https://github.com/bat/bat/issues/5#issuecomment-54267956
[3] https://twiki.cern.ch/twiki/bin/view/CMSPublic/SWGuideCreatingNewProducts