Modify PECInputData to own all trees it reads from a PEC file
The motivation behind this change is that dependant plugins cannot read
entries from trees using only pointers to the trees; they also need to
know the index of the current event. The index is an internal property
of PECInputData and should not be exposed to the world.

Solving this annoyance changes logic of PECInputData significantly,
through in the right direction of an "input file handler". Access to
trees is implemented with a map, which is not very nice for reading of
every event. However, no visible degradation in performance is expected.
Still, it would be good to run profiling at some point.