Make method PECInputData::LoadTree constant
Processor must not give a non-const pointer to a plugin to other plugins
(so that they cannot reset its master, call ProcessEvent, etc). But they
must be capable of calling LoadTree. In fact, constantness in Plugin is
mostly intended to discriminate between what Processor and other plugins
can do to it.