Update trigger filtering
Classes for trigger filtering are converted to plugins and modified
heavily to align them with the new design of the framework.

Currently, modifications lead to a crash when accessing pattern of the
MC trigger in PECTriggerFilterMC::BeginRun. Needs to be investigated.

In addition, compilation errors are triggered in test/multithread.cpp.
They will be solved when the old TriggerSelectionInterface is eliminated
completely.