Fix clonning of various readers of PEC format
Default copy constructors were used, which is not correct because they
would copy addresses of buffers to read branches from trees (this nasty
feature of ROOT that demands providing a pointer to pointer to the
buffer). This would not trigger a problem because plugins were clonned
using the initializing constructor, not the copy one. But this solution
is not correct since any customisation (e.g., non-standard names for
plugins providing inputs) would be lost.

This fixes #41.