Add a plugin that fills jet histograms need to compute b-tagging effs
This is on the boundary with analysis-specific plugins, but since
everybody needs to compute b-tagging efficiencies in simulation, it
would be nice to have this helper plugin in the distribution.

The added implementation is not complete yet. Need to add access to
event weights (at least pile-up and nominal GEN-level).