Remove TDirectory::Append in TFileService
This undoes commit 20b183603bbd0c5ed97a9419408d857e5ddfb407. With
TDirectory::Append was getting a crash at the end of a job if
TFileService is used to create trees. Don't really understand the
reason of the problem, but it was supposed to work without this line in
the first place anyway.