Update b-tagging efficiencies and scale factors to 76X
There is a problem in the unit test for scale factors for light-flavour
jets with pt = 1 TeV: the scale factor is zero. Will need to
investigate.

File with inputs for "CSV reweighting" is removed: in 76X these inputs
are included in the regular CSV files. (Would need to update
BTagSFService though, see issue #50).