Allow application of T1 MET corrections directly to raw MET
It has been validated with data files that this commit results in the
same values for MET as the subtraction/correction scheme for majority of
events. For few events where the results are different, the difference
is assumed to be caused by the non-trivial jet selection applied by the
PAT MET tool.