Include electron scale factors
Use scale factors presented in this talk [1]. Their uncertainties cover
the degradation of trigger efficiency in runs affected by
misreconstruction of the online beam spot [2], and thus this portion of
data is now included. As a consequence, pile-up profile in the electron
channel is also updated, and now it's virtually indistinguishable from
the one in the muon channel.
[1] https://indico.cern.ch/event/491510/contributions/2246475/attachments/1310813/1962318/2016.07.18_EGM_Electron-trigger-SF_v2.pdf
[2] https://hypernews.cern.ch/HyperNews/CMS/get/physics-validation/2556.html