Specify dependencies in libmensura.so
With this fix applications that link against this library need not
specify all its dependencies like ROOT and boost_filesystem.

However, the path to Boost libraries is included as an rpath, and,
apparently, it will not be picked up by the applications. The short-term
solution is to build them with -Wl,-rpath=$(BOOST_ROOT)/lib. If there is
no other work-around, should better switch to using LD_LIBRARY_PATH in
future.