Add Doxygen documentation
Generated from commit 5fa277b8d95277afea3c683d2bd4f4ed285c8a30.

Currently the workflow uses two clones of the mensura repository. In one
the branch gh-pages is checked out, and doxygen is run from within it.
In the other the master branch is checked out, and it is given as the
input to the the doxygen. The relative path to this second repository is
hard-coded in Doxyfile.

Since doxygen creates some files whose names start with underscores,
manually create file .nojekyll in the doxygen/html/ directory.