Update b-tagging scale factors
It is not clear when and why the change in the scale factors was
introduced. Only the "ttbar" measurement is affected, while data for
"incl" and "iterativefit" are the same as before. It might be that the
previous version (v2) was the preliminary one.

Drop scale factors for separate data taking periods since they are not
used.