Add a warning in GenWeightSyst
It is printed when user has requested normalization by mean weights, but
values for the current dataset have not been found.

In the past it was desirable to avoid such a warning because not
providing mean weights was a way to switch off the corresponding
systematics on a per-dataset basis. This is not needed anymore after a
constructor from a JSON file has been introduced.