Enable built-in thread awareness in ROOT
Observed occasional segmentation faults when running in multiple threads
and recorrecting jets. The exact origin of the problem is not known, but
it might be related to JIT compilation in TFormula. Instead of adding
another ROOTLock, tried newly introduced ROOT::EnableThreadSafety(),
which is supposed to make ROOT (partially) thread-safe. There is no
documentation of what exactly it protects [1], however. Nonetheless,
this change seems to fix the crash. ROOTLock might not be needed
anymore.

[1] https://root-forum.cern.ch/t/problem-with-std-thread-in-root/25600/8