Refactor Dataset removing File subclass
Assign the normalization parameters (cross section, number of events,
and mean weight) to Dataset instead of individual instances of
Dataset::File included in a data set. Now a single data set can only
describe a single physics process, which was the case de facto anyway.

This change is not bacward-compatible and will break user code.